fix: remove shared package and fix docker build configuration
- Remove unused @turnos/shared package directory entirely
- Fix Dockerfile WORKDIR consistency (use /app instead of /workspace)
- Update docker-compose volumes to match Dockerfile paths
- Clean up package.json and tsconfig references to @turnos/shared
- Remove stale dist/ volume mounts from compose (NestJS --watch recompiles)
- Both consumer and producer now have self-contained domain entities
- All containers now start successfully with proper module resolution

Services tested:
✅ backend-producer (port 3000, Swagger available)
✅ backend-consumer (listening on turnos_queue, scheduler running)
✅ frontend-app (port 3001)
✅ RabbitMQ (port 5672)
✅ MongoDB (port 27017)
