+++ backend/consumer/Dockerfile
@@ -1,27 +1,9 @@
-# ⚕️ HUMAN CHECK - Dockerfile del Consumer
-# Build context = ./backend (para acceder a shared/)
-# Asegurarse de usar una versión LTS de Node.js compartida con el Producer
-
 # --- Stage: development ---
-# ⚕️ HUMAN CHECK - Stage de desarrollo para uso con docker-compose (hot-reload)
 FROM node:20-slim AS development
 
-# Estructura /workspace/consumer + /workspace/shared replica backend/ local
-# para que "file:../shared" en package.json resuelva correctamente
-WORKDIR /workspace
+WORKDIR /app
 
-# Copiar shared primero (dependencia local del consumer)
-COPY shared/ ./shared/
-WORKDIR /workspace/shared
-RUN npm run build
-
-# Volver a /workspace antes de copiar consumer
-WORKDIR /workspace
-
-# Instalar dependencias del consumer
-COPY consumer/package.json ./consumer/
-COPY consumer/package-lock.json ./consumer/
-WORKDIR /workspace/consumer
+COPY consumer/package*.json ./
 RUN npm install
 
 # Copiar código fuente (en dev, volumes sobreescriben esto)
@@ -32,18 +14,9 @@ CMD ["npm", "run", "start:dev"]
 # --- Stage: build ---
 FROM node:20-slim AS build
 
-WORKDIR /workspace
+WORKDIR /app
 
-COPY shared/ ./shared/
-WORKDIR /workspace/shared
-RUN npm run build
-
-# Volver a /workspace antes de copiar consumer
-WORKDIR /workspace
-
-COPY consumer/package.json ./consumer/
-COPY consumer/package-lock.json ./consumer/
-WORKDIR /workspace/consumer
+COPY consumer/package*.json ./
 RUN npm install
 
 COPY consumer/ .
@@ -52,25 +25,11 @@ RUN npm run build
 # --- Stage: production ---
 FROM node:20-slim AS production
 
-WORKDIR /workspace
+WORKDIR /app
 
-# ⚕️ HUMAN CHECK - Compilar shared primero, luego copiar dist compilado
-COPY shared/ ./shared/
-WORKDIR /workspace/shared
-RUN npm run build
+COPY consumer/package*.json ./
+RUN npm install --omit=dev
 
-# Volver a /workspace antes de copiar consumer
-WORKDIR /workspace
-
-COPY consumer/package.json ./consumer/
-COPY consumer/package-lock.json ./consumer/
-WORKDIR /workspace/consumer
-RUN npm install --only=production
-
-COPY --from=build /workspace/consumer/dist ./dist
-
-# ⚕️ HUMAN CHECK - Variables de Entorno
-# El Consumer no expone puerto HTTP necesariamente, pero se puede monitorear
-# EXPOSE 3000
+COPY --from=build /app/dist ./dist
 
 CMD ["node", "dist/main.js"]
