+++ backend/consumer/package.json
@@ -17,7 +17,6 @@
         "test:cov": "jest --coverage"
     },
     "dependencies": {
-        "@turnos/shared": "file:../shared",
         "@nestjs/common": "^10.0.0",
         "@nestjs/config": "^3.0.0",
         "@nestjs/core": "^10.0.0",
