+++ backend/producer/Dockerfile
@@ -3,19 +3,11 @@
 # Asegurarse de usar una versión LTS de Node.js
 
 # --- Stage: development ---
-# ⚕️ HUMAN CHECK - Stage de desarrollo para uso con docker-compose (hot-reload)
 FROM node:20-slim AS development
 
-# Estructura /workspace/producer + /workspace/shared replica backend/ local
-# para que "file:../shared" en package.json resuelva correctamente
-WORKDIR /workspace
+WORKDIR /app
 
-# Copiar shared primero (dependencia local del producer)
-COPY shared/ ./shared/
-
-# Instalar dependencias del producer
-COPY producer/package*.json ./producer/
-WORKDIR /workspace/producer
+COPY producer/package*.json ./
 RUN npm install --legacy-peer-deps
 
 # Copiar código fuente (en dev, volumes sobreescriben esto)
@@ -28,11 +20,9 @@ CMD ["npm", "run", "start:dev"]
 # --- Stage: build ---
 FROM node:20-slim AS build
 
-WORKDIR /workspace
+WORKDIR /app
 
-COPY shared/ ./shared/
-COPY producer/package*.json ./producer/
-WORKDIR /workspace/producer
+COPY producer/package*.json ./
 RUN npm install --legacy-peer-deps
 
 COPY producer/ .
@@ -41,7 +31,7 @@ RUN npm run build
 # --- Stage: production ---
 FROM node:20-slim AS production
 
-WORKDIR /workspace
+WORKDIR /app
 
 # ⚕️ HUMAN CHECK - En producción, shared se copia para resolver el symlink file:../shared
 COPY shared/ ./shared/
