+++ backend/producer/tsconfig.json
@@ -10,10 +10,6 @@
         "sourceMap": true,
         "outDir": "./dist",
         "baseUrl": "./",
-        "paths": {
-            "@turnos/shared": ["../shared/src"],
-            "@turnos/shared/*": ["../shared/src/*"]
-        },
         "incremental": true,
         "skipLibCheck": true,
         "strictNullChecks": false,
