--- backend/shared/package.json
@@ -1,12 +1,0 @@
-{
-    "name": "@turnos/shared",
-    "version": "1.0.0",
-    "description": "Tipos y entidades de dominio compartidas entre Producer y Consumer",
-    "private": true,
-    "main": "src/index.ts",
-    "types": "src/index.ts",
-    "license": "UNLICENSED",
-    "scripts": {
-        "build": "echo 'Shared package - no build required'"
-    }
-}
