--- backend/shared/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "declaration": true,
-        "removeComments": true,
-        "target": "es2017",
-        "sourceMap": true,
-        "outDir": "./dist",
-        "baseUrl": "./",
-        "incremental": true,
-        "skipLibCheck": true,
-        "strict": false
-    },
-    "include": ["src/**/*"]
-}
